@jreback 2021-12-07
> pandas/core/internals/managers.py:1090: error: Argument 1 to "_iset_single" of "BlockManager" has incompatible type "Union[int, slice, ndarray[Any, Any]]"; expected "int"  [arg-type]

looks good otherwise