@jbrockmendel 2021-12-07
still looking at asvs; the full run showed some slowdowns that don't make sense to me so i need to do some digging